+++ CHANGELOG.md
@@ -8,6 +8,7 @@
       * Require `PayPalCheckoutRequest.userAuthenticationEmail` for App Switch support
       * **Note:** This feature is currently in beta and may change or be removed in future releases.
   * Send `payPalContextId` in `PayPalLauncher.handleReturnToApp` related events.
+  * Fix a bug to correctly log app switch flow analytics.
 
 ## 5.6.0 (2025-02-05)